Fixed former errors, and completed Step 1 and Step 3
had a disastrous time loading libraries, but i think i fixed the issue. it was mainly an issue on my machine. oh well, live and learn. libraries are loading good now, and step1 is complete, and step3 is good too. Not sure how necessary step 2 is but thats the next thing to tackle